refactor(frontend): create reusable getSwapExplorerLink util
