+++ packages/frontend/src/hooks/useSwapToast.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { showToast, Toast, updateToast } from '@sifi/shared-ui';
 import { JumpStatus } from '@sifi/sdk';
 import { usePublicClient } from 'wagmi';
-import { getEvmTxUrl, getViemErrorMessage } from 'src/utils';
+import { getSwapExplorerLink, getViemErrorMessage } from 'src/utils';
 import { useSpaceTravel } from 'src/providers/SpaceTravelProvider';
 import { useSwapFormValues } from './useSwapFormValues';
 import { useRefetchBalances } from './useRefetchBalances';
@@ -22,15 +22,6 @@ const useSwapToast = () => {
   const isJump = fromChain.id !== toChain.id;
   const [swapToastId, setSwapToastId] = useState<string | number | null>(null);
   const [toastLink, setToastLink] = useState<{ text: string; href: string } | undefined>(undefined);
-
-  // TODO: Create a reusable utility function for this
-  const getExplorerLink = (hash: `0x${string}`): string | undefined => {
-    if (isJump) {
-      return `https://layerzeroscan.com/tx/${hash}`;
-    } else {
-      return getEvmTxUrl(fromChain, hash);
-    }
-  };
 
   const updateJumpToast = (status: JumpStatus) => {
     let toastContent = {};
@@ -87,7 +78,7 @@ const useSwapToast = () => {
   };
 
   const showSwapToast = async ({ hash }: { hash: `0x${string}` }) => {
-    const explorerLink = getExplorerLink(hash);
+    const explorerLink = getSwapExplorerLink(fromChain, toChain, hash);
     const newToastLink = explorerLink ? { text: 'View Jump', href: explorerLink } : undefined;
     // Need to store this in state so updateJumpToast can access it
     setToastLink(newToastLink);
@@ -116,6 +107,7 @@ const useSwapToast = () => {
         setThrottle(0.01);
         updateToast(swapToastId, {
           render: <Toast text="Swap completed." type="success" link={newToastLink} />,
+
           closeOnClick: true,
         });
       }
